Re-Add Solidus/TaxCategoryDeprecated cop
This cop was removed earlier as it was giving a lot of false positives.

We have re-added this cop again but now with a warning offense so that it doesn't raise any errors for this cop
but warns the user to have a look at certain places in the code where something might be wrong.