Fix overridden add_offense method
In d525445, we updated the `add_offense` method to let it work with
older ruby versions. Everything worked fine until we started passing
other params to the method, such as the severity, which doesn't work
in Ruby 2.7 because the behavior is incorrect. Our way of calling
the super method was introduced in Ruby 2.7, so we need to start using
the new way from the correct version.